Revise Chapter 1, Revise README.md
This marks the beginning of my revision for this repository. I added some new Chapter 1 notes, and am working on changing the overall file structure so code/notes are grouped by chapter.
I adjusted the README to more accurately reflect my current goals for this repository. Instead of it just being a history of my progress for PY4E, I want it to be suitable for those who would use this repository as a learning aide for PY4E.
Hopefully whenever I eventually force my close friends and family to learn Python, this aide will make the process much more enjoyable for them.