Revise Chapter 1 Notes, Revise Chapter 2 Exercises
I decided to make Chapter 1 into an introduction/set-up guide that would show beginners how to access or create a Linux CLI environment on their machine.
The guide also helped them initialize the environment for Python development, and gives an extremely basic 'Hello World' tutorial to ensure everything is up and running.
I also decided to change the directory containing 'hello_world.py' to 'exercises'. This implies that the structure of each Chapter directory should contain notes.md and a directory for exercises.