Revise README.md, Chapter 3 notes.md, Chapter 2 notes.md
Here's the revision I mentioned in the last commit. I changed the README.md to reflect the addition of Chapter 3.
I also made a small tweak in the naming of the sections in Chapter 3 notes.md and Chapter 2 notes.md that discuss the chapter exercises.
Now it's time to go eat a chicken salad sandwich.