Revise Chapter 3 exercises 1, 2, and answers
When I checked on PY4E, I realized that I didn't get the expected output correct for exercise 2, so I made that adjustment.
Also, I added "" to denote when an expected output is supposed to be a string printed to the console. This change was made on all 4 files in the Chapter 3 exercises directory.