Revise Chapter names, Chapter exercise file names, and Chapter 3.
After going through lesson 4 on PY4E, I realized that the 'first' chapter in my repository was not counted in the chapter count, so I relabled it to Chapter 0 and made the appropriate changes for following chapters.
I added new exercises and notes to Chapter 3, which can be found in Chapter_3/exercises/ and Chapter_3/notes.md respectively.
I also made some clarifying changes to Chapter 2 notes.md, specifically explaining  how I represent Python code for the remainder of the course.
Now yet again, it's time to eat a chicken salad sandwich for lunch.