Revise Chapter 5
Well, it's been a while. I guess that's what happens when life gets unpredictable. This is the reworking of Chapter 5 (or Chapter 6 in PY4E, including the Installation Chapter)
From now on, with life being so tumultuous, I'm going to shift more towards quick notes/highlights and away from lenghier writing (at least for the time being)
I still believe that this repo will still be able to serve as a guide for others, but the remaining chapters will begin to look more like cheat-sheets and less like additional commentary on the course.
When life slows down a little bit, I will be eager to revise the following chapters into the current 'additional commentary' format.
This commit comment is written out just as much for me as for anyone else reading it.