Add Chapter 16, Revise README.md Added Chapter 16, which focused on putting all previous concepts together to form several data mining applications. I also made the necessary revisions to the README to reflect the a ddition of Chapter 16.
This is the end of the course, so no more additional chapters will be added. Sooner (or later), I will flesh out the applications discussed in the final lectures so they are functioning within this repository.
There is also a need to adjust the code to fit the Python Enhancement Proposals (PEP) 8 Style Guide. That will also come sooner (or later).
Until then, I hope you've enjoyed perusing this comprehensive guide->turned->cliff notes.